@@ -1,8 +1,10 @@
-import React from 'react'
+import React, {Component} from 'react'
 import { Link } from 'react-router'
 
-export default React.createClass({
+class NavLink extends React.Component{
   render() {
     return <Link {...this.props} activeClassName="active"/>
   }
-})
+}
+
+export default NavLink;
