@@ -3,7 +3,7 @@ import { render } from 'react-dom'
 import { Router, Route, browserHistory,IndexRoute } from 'react-router'
 import App from './modules/App'
 import ES6 from './modules/ES6'
-import Repos from './modules/Repos'
+import ES7 from './modules/ES7'
 import Repo from './modules/Repo'
 import NotFound from './modules/NotFound'
 import NewCapability from './modules/newCapability'
@@ -17,8 +17,8 @@ render((
       <Route path="/es6" component={ES6}>
         <Route path="/es6/:newCapability" component={NewCapability}/>
       </Route>
-      <Route path="/repos" component={Repos}>
-        <Route path="/repos/:userName/:repoName" component={Repo}/>
+      <Route path="/repos" component={ES7}>
+        <Route path="/es7/:newCapability" component={NewCapability}/>
       </Route>
 
     </Route>
